Fixed
Turns out the super optimistic 2200% performance improvement is just from the function wiping the array after the first run. So the rest of hte 9999 times are running against an empty array.
With a array.slice() fix, the old version is actually 500% faster.
Optimized a bit, now the old version is still ~70% faster. But less line count = more gooder.